services/zero-trust-security/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
pyjwt==2.8.0
cryptography==41.0.7
prometheus-client==0.19.0
